Data shown in shops are now from firestore
new error emerged:
Setting a timer for a long period of time, i.e. multiple minutes, is a performance and correctness issue on Android as it keeps the timer module awake, and timers can only be called when the app is in the foreground.
This error seems to be from firebase itself or at least the implementation of read function